// src/types/onboarding.ts
export interface I_Register {
    business_name: string;
    business_email: string;
    business_phone_num: string;
    business_category: string;
    acct_number: string;
    logo: string;
    house_num: string;
    street: string;
    city: string;
    state: string;
    contact_name: string;
    contact_phone: string;
    contact_email: string;
    password: string;
  }